Error handling and security roles
added error handling to all the controllers and services excl. the roles and user. i still need to do that. also i created security around my endpoints. i also tested in postman if USER and ADMIN are working correctly. As of today it is working and all the pathings now work only when either ADMIN (all privileges) or User (limited privileges).